requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
rich>=13.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
rapidfuzz>=3.0.0

# PDF extraction
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# LLM clients
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
